query_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
cassandra-driver==3.29.1
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1